move landing page component, create method to enter main site
